+++ examples/images.nim
@@ -7,9 +7,7 @@ import pkg/bard
 
 dotenv.load()
 
-let ai = waitFor newBardAi(
-  psid = getEnv("bard_psid"),
-  psidts = getEnv("bard_psidts")
-)
+let ai = waitFor newBardAi getEnv "bard_cookies"
 
-echo waitFor(ai.prompt "show me cats images").text
+for img in waitFor(ai.prompt "show me cats images").images:
+  echo img[]