+++ examples/single.nim
@@ -7,9 +7,6 @@ import pkg/bard
 
 dotenv.load()
 
-let ai = waitFor newBardAi(
-  psid = getEnv("bard_psid"),
-  psidts = getEnv("bard_psidts")
-)
+let ai = waitFor newBardAi getEnv "bard_cookies"
 
 echo waitFor(ai.prompt "Tell me an Asian traditional history in ten words").text