+++ src/bard.nim
@@ -2,10 +2,13 @@ import std/asyncdispatch
 from std/httpclient import newAsyncHttpClient, newHttpHeaders, postContent, close,
                             HttpHeaders, getContent, HttpRequestError
 from std/strformat import fmt
-from std/strutils import split, replace
+from std/strutils import split, replace, join
 from std/json import `$`, `%*`, newJNull, parseJson, `{}`, len, getStr, JsonNode,
                       items, JNull, JsonParsingError
+export json.items # Why it doesn't compiles without this?
+
 from std/uri import `$`, encodeQuery, Uri
+from std/sugar import collect
 
 from pkg/util/forStr import between
 
@@ -18,6 +21,7 @@ type
     headers: HttpHeaders
     snlm0e: string
     reqId: int
+    server: string
   BardAiChat* = ref object
     ai*: BardAi
     conversationId: string
@@ -26,6 +30,10 @@ type
   BardAiResponse* = ref object
     text*: string
     images*: seq[BardAiImage]
+    drafts*: seq[BardAiResponseDraft]
+    relatedSearches*: seq[string] ## "Google it" button
+  BardAiResponseDraft* = tuple
+    id, text: string
   BardAiImage* = ref object
     tag*: string
     url*: string
@@ -63,8 +71,15 @@ func nextReqId(self): int =
   else:
     self.reqId + sum
 
-proc newBardAi*(psid, psidts: string): Future[BardAi] {.async.} =
+const bardServers* = [
+  "boq_assistant-bard-web-server_20230419.00_p1",
+  "boq_assistant-bard-web-server_20230711.08_p0",
+]
+
+proc newBardAi*(cookies: string; server = bardServers[0]): Future[BardAi] {.async.} =
   ## Creates new Bard AI object
+  ## 
+  ## Required cookies: `__Secure-1PSID` and `__Secure-1PSIDTS`
   new result
   result.headers = newHttpHeaders({
     "Host": hostUrl,
@@ -73,12 +88,9 @@ proc newBardAi*(psid, psidts: string): Future[BardAi] {.async.} =
     "Content-Type": "application/x-www-form-urlencoded;charset=UTF-8",
     "Origin": baseUrl,
     "Referer": baseUrl & "/",
-    "Cookie": fmt"__Secure-1PSID={psid}; __Secure-1PSIDTS={psidts}",
-    # "Sec-Fetch-Dest": "empty",
-    # "Sec-Fetch-Mode": "cors",
-    # "Sec-Fetch-Site": "same-origin",
-    # "Sec-GPC": "1",
+    "Cookie": cookies
   })
+  result.server = server
   result.reqId = nextReqId result
   await result.getSNlM0e
   
@@ -91,6 +103,7 @@ proc newBardAiChat*(self): BardAiChat =
 template headers(chat): BardAi.headers = chat.ai.headers
 template reqId(chat): BardAi.reqId = chat.ai.reqId
 template snlm0e(chat): BardAi.snlm0e = chat.ai.snlm0e
+template server(chat): BardAi.server = chat.ai.server
 
 func buildQuery(self: BardAi or var BardAiChat; prompt: string): string =
   ## Creates the Bard's batchexecute query body
@@ -120,7 +133,7 @@ func buildUrl(self: BardAi or var BardAiChat): Uri =
   result.hostname = hostUrl
   result.scheme = "https"
   result.query = encodeQuery({
-    "bl": "boq_assistant-bard-web-server_20230711.08_p0",
+    "bl": self.server,
     "_reqID": $self.reqId,
   })
 
@@ -128,8 +141,7 @@ proc getRespJson(resp: string): JsonNode =
   try:
     let json = resp.split("\n")[2].parseJson{0}{2}
     if not json.isNil:
-      return json.getStr.parseJson
-    
+      return parseJson getStr json
   except JsonParsingError:
     discard
   raise newException(BardUnrecognizedResp,
@@ -167,9 +179,19 @@ proc prompt*(self: BardAi or var BardAiChat; text: string): Future[BardAiRespons
     echo getCurrentExceptionMsg()
     raise newException(BardExpiredSession, "Your Google account session was expired")
   let bardData = resp{4, 0}
+  
   new result
   result.images = extractImages bardData{4}
   result.text = bardData{1, 0}.getStr.applyImages result.images
+
+  for draft in resp{4}:
+    result.drafts.add BardAiResponseDraft (
+      id: draft{0}.getStr,
+      text: collect(for txt in draft{1}: getStr txt).join "\l"
+    )
+  
+  for suggestion in resp{2}:
+    result.relatedSearches.add getStr suggestion{0}
 
   when self is BardAiChat:
     self.conversationId = resp{1}{0}.getStr
@@ -177,10 +199,7 @@ proc prompt*(self: BardAi or var BardAiChat; text: string): Future[BardAiRespons
     self.choiceId = resp{4}{0}{0}.getStr
 
 when isMainModule:
-  let ai = waitFor newBardAi(
-    psid: "",
-    psidts: ""
-  )
+  let ai = waitFor newBardAi("__Secure-1PSIDTS=sidts-CjEBSAxbGT-Wp1R1dKZoAfEdCPDi4tRLF0ITxS7KH6gWK1XAOB2_5xz0XL5zHIPBhwG5EAA; __Secure-1PSID=ZwjrVdNd2_lQO0-AseXdwV-CpwuumXtwHhWHiyyRWWi4xtaOdXVU_grpOx4DLpAkNy_nJw.")
   var chat = ai.newBardAiChat
   # echo waitFor ai.prompt "Tell me an Asian traditional history in ten words"
   echo waitFor(chat.prompt "my name is jeff")[]